fix memory leak issue and skip downloaded files

+++ browser.py
@@ -9,10 +9,11 @@ import os
 from datetime import datetime, timedelta
 import sys
 
-begin = 2015
+begin = 2003
 end = 2015
 
 now = datetime.now()
+
 
 URL = 'http://fhy.wra.gov.tw/ReservoirPage_2011/StorageCapacity.aspx'
 br = mechanize.Browser()
@@ -21,14 +22,14 @@ br.addheaders = [('User-agent', 'Mozilla/5.0 (X11; U; Linux i686; en-US; rv:1.9.
 """ Get days of every month """
 for reservior in {'主要水庫', '所有水庫', '水庫及攔河堰'}:
 	for year in range(begin, end + 1):
-		for month in range(2, 13):
+		for month in range(1, 13):
 			days = calendar.monthrange(year, month)[1]
 
 			for day in range(1, days + 1):
 				""" initialize list and dictionary """
 				info = []
 				content = {}
-				
+
 				""" it will delay 1~2 days for data update, so will fetch (now - 2) days """
 				if datetime(year, month, day) > now - timedelta(days=2):
 					print "date upate to", year, month, (day - 1)
@@ -43,6 +44,12 @@ for reservior in {'主要水庫', '所有水庫', '水庫及攔河堰'}:
 					print "create directory"
 					os.makedirs(directory)
 				filename = str(month) + '-' + str(day) + '.json'
+				
+				""" skip file if it exist """
+				if os.path.isfile(directory + '/' + filename):
+					print "skip ", str(year), str(month), str(day)
+					continue
+
 				f = open(directory + '/' + filename, 'w')
 
 				""" Generate data for post form """
@@ -63,7 +70,10 @@ for reservior in {'主要水庫', '所有水庫', '水庫及攔河堰'}:
 
 				response = br.submit()
 				page =  br.response().get_data()
-
+				
+				""" clear history data of mechanize to release memory""" 
+				r.close()
+				br.clear_history()
 
 				""" Read response """
 				soup = BeautifulSoup(page)
@@ -81,32 +91,33 @@ for reservior in {'主要水庫', '所有水庫', '水庫及攔河堰'}:
 					m = re.match(r'(<tr>|<tr class="alternate">)\s*<td>(.*)</td><td align="right">(.*)</td><td>\s*(.*)<br/>\s*(.*)\s*</td><td align="right">(.*)</td><td align="right">(.*)</td><td align="right">(.*)</td><td align="right">(.*)</td><td>(.*)</td><td>(.*)</td><td align="right">(.*)</td><td align="right">(.*)</td><td align="right">(.*)</td>', str(element), re.M)
 
 					if m:
-						#print repr(m.group(4))
-						#print m.group(2), m.group(3), m.group(4), m.group(5), m.group(6) 
 						#print m.group(2), m.group(3), m.group(4), m.group(5), m.group(6), m.group(7), m.group(8), m.group(9), m.group(10), m.group(11), m.group(12), m.group(13), m.group(14)
-						content = {
-					"name": m.group(2),
-					"daily": {
-						"capacity": m.group(3),
-						"start_time": m.group(4),
-						"end_time": m.group(5),
-						"rain": m.group(6),
-						"in_water": m.group(7),
-						"out_water": m.group(8),
-						"diff": m.group(9),
-						"nuclear": m.group(10),
-						},
-					"in_time": {
-						"time": m.group(11),
-						"height": m.group(12),
-						"capacity": m.group(13),
-						"now_cap_percent": m.group(14),
-						}
+						content = 
+							{
+							"name": m.group(2),
+							"daily": 
+								{
+								"capacity": m.group(3),
+								"start_time": m.group(4),
+								"end_time": m.group(5),
+								"rain": m.group(6),
+								"in_water": m.group(7),
+								"out_water": m.group(8),
+								"diff": m.group(9),
+								"nuclear": m.group(10),
+								},
+							"in_time": 
+								{
+								"time": m.group(11),
+								"height": m.group(12),
+								"capacity": m.group(13),
+								"now_cap_percent": m.group(14),
+								}
 							}
 						info.append(content)
 
 				f.write(json.dumps(info, ensure_ascii=False))
-				f.close
+				f.close()
